+++ jbt/src/test/java/org/hibernate/tool/orm/jbt/api/wrp/PersistentClassWrapperTest.java
@@ -30,7 +30,7 @@ import org.hibernate.tool.orm.jbt.internal.factory.PropertyWrapperFactory;
 import org.hibernate.tool.orm.jbt.internal.factory.TableWrapperFactory;
 import org.hibernate.tool.orm.jbt.internal.factory.ValueWrapperFactory;
 import org.hibernate.tool.orm.jbt.internal.util.DummyMetadataBuildingContext;
-import org.hibernate.tool.orm.jbt.util.SpecialRootClass;
+import org.hibernate.tool.orm.jbt.internal.util.SpecialRootClass;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
